Autonomous Wheel Colliders
Wheel Colliders now work autonomously.
This means we don't need the basic meshes to provide wheel orientation and we don't need any sphere colliders for the wheels to work properly.
We also only need one mesh for all 4 wheels.
This comes at the expense of 1 if-condition per fixed update (frame), since we have to somehow check which wheel mesh is left and needs to be 180° rotated.